fix: /user/[userId]/_components 폴더 만들어서 구조 수정
