+++ src/pages/user/[userId]/_components/ActivitySection.tsx
@@ -1,0 +1,37 @@
+import { useParams } from 'next/navigation';
+import useUserProfile from '@/models/user/profile/useUserProfile';
+import ActivityCard from '@/shared/components/ActivityCard/ActivityCard';
+
+const ActivitySection = () => {
+  const params = useParams();
+  const { data: user } = useUserProfile(Number(params?.userId));
+
+  return (
+    <div className="mt-[50px] space-y-[30px] xl:mt-0">
+      <div>활동 내역</div>
+      <div className="flex space-x-2.5 xl:space-x-5">
+        <div className="w-full">
+          <ActivityCard
+            status="averageLeft"
+            conScore={user?.data.averageRating}
+          />
+        </div>
+        <div className="w-full">
+          <ActivityCard
+            status="reviewsLeft"
+            conScore={user?.data.reviewCount}
+          />
+        </div>
+        <div className="w-full">
+          <ActivityCard
+            status="interest"
+            text={user?.data.mostFavoriteCategory?.name}
+            color="#23b581"
+          />
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default ActivitySection;
